feat(connection): added ability to get refresh token from juncture
This method is discouraged though, since if users try to manage the refresh token and juncture tries to manage the refresh token, this can lead to conflicts. It is preferred to get the access_token from juncture instead, and let juncture handle the refresh_token logic. Only use this method if you decide not to use the juncture integration helpers at all